Distribution changes
* Created `flag.js` controller as per sizzle and moved the export there
* Created `vote.js` controller base
* Move/remove last commit stuff out that was added in.
* Renamed `setFlaggedListInModel` to `setFlaggedListToModel` ... this still could use some advice on exact name that is wanted for the identifier.
* NOTE'd a possible extra dir traversal that shouldn't be needed as compared to below requires... this is beyond this PR and will retest/cleanup in another unrelated commit so just noting it

Applies to #643
