+++ controllers/vote.js
@@ -1,0 +1,8 @@
+'use strict';
+
+// Define some pseudo module globals
+var isPro = require('../libs/debug').isPro;
+var isDev = require('../libs/debug').isDev;
+var isDbg = require('../libs/debug').isDbg;
+
+//
